moved new code to original scrape route
